+++ .config/dotnet-tools.json
@@ -1,0 +1,12 @@
+{
+  "version": 1,
+  "isRoot": true,
+  "tools": {
+    "csharpier": {
+      "version": "0.27.3",
+      "commands": [
+        "dotnet-csharpier"
+      ]
+    }
+  }
+}
