+++ WebApplication1/Hello.cs
@@ -4,15 +4,12 @@ namespace WebApplication1
     {
         public string Name { get; set; }
         public string Greeting { get; set; }
-
-
         public Hello() {
             
             Name = "Buddy, I am free! You have released me from my prison due to your mistake!";
             Greeting = "Hello World";
 
         }
-
 
     }
 }
